Reorganize betty.datas
