--- betty/datas/human_facing_plugin_definition.py
@@ -1,40 +1,0 @@
-"""
-Human-facing plugin definition data.
-"""
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Any
-
-from betty.attrs.localizable import new_localizable_attr
-from betty.attrs.optional import OptionalAttr
-from betty.datas.plugin_definition import PluginDefinitionData
-from betty.localizables.gettext import _
-from betty.plugin import PluginDefinition
-
-if TYPE_CHECKING:
-    from betty.localizable import ResolvableLocalizable
-
-
-class HumanFacingPluginDefinitionData[
-    PluginDefinitionT: PluginDefinition = PluginDefinition
-](PluginDefinitionData[PluginDefinitionT]):
-    """
-    Configure a :py:class:`betty.definition.human_facing.HumanFacingDefinition`.
-
-    .. data:: betty.datas.human_facing_plugin_definition:HumanFacingPluginDefinitionData
-    """
-
-    label = new_localizable_attr(label=_("Label"))
-    description = OptionalAttr(new_localizable_attr(label=_("Description")))
-
-    def __init__(
-        self,
-        *,
-        label: ResolvableLocalizable,
-        description: ResolvableLocalizable | None = None,
-        **kwargs: Any,
-    ):
-        super().__init__(**kwargs)
-        self.label = label
-        self.description = description
