+++ betty/datas/plugin/__init__.py
@@ -1,0 +1,3 @@
+"""
+Reusable definitions for plugin-related data.
+"""
