+++ betty/datas/plugin/definition/copyright_notice.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Any, final, override
 from betty.attrs.localizable import new_localizable_attr
 from betty.copyright_notice import CopyrightNotice, CopyrightNoticeDefinition
 from betty.datas.aggregate.record.object import ObjectDefinition
-from betty.datas.human_facing_plugin_definition import HumanFacingPluginDefinitionData
+from betty.datas.plugin.definition.human_facing import HumanFacingPluginDefinitionData
 from betty.localizables.gettext import _
 from betty.sample import Sample
