+++ betty/datas/plugin/definition/ordered.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Any
 
 from betty.attrs.owner import CollectionOwnerAttr
 from betty.datas.aggregate.collection.sequence import SequenceDefinition
-from betty.datas.plugin_definition import PluginDefinitionData
+from betty.datas.plugin.definition import PluginDefinitionData
 from betty.localizables.gettext import _
 from betty.machine_name import MachineName
 from betty.plugin.ordered import OrderedPluginDefinition
