+++ betty/datas/plugin/manufacturer/__init__.py
@@ -1,0 +1,3 @@
+"""
+Reusable data for plugin manufacturers.
+"""
