+++ betty/datas/plugin/manufacturer/sequence.py
