+++ betty/tests/attrs/test_plugin_definitions.py
@@ -4,7 +4,7 @@ from typing import override
 from betty.attrs.plugin_definitions import new_plugin_definition_datas_attr
 from betty.data import Data
 from betty.datas.aggregate.record.object import ObjectDefinition
-from betty.datas.plugin_definition import PluginDefinitionData
+from betty.datas.plugin.definition import PluginDefinitionData
 from betty.prop import HasProps
 from betty.test_utils.plugin import DummyPluginDefinition
