+++ betty/tests/datas/plugin/definition/test_copyright_notice.py
@@ -1,13 +1,13 @@
 from __future__ import annotations
 
-from betty.datas.license_definition import LicenseDefinitionData
+from betty.datas.plugin.definition.copyright_notice import CopyrightNoticeDefinitionData
 from betty.localizables.plain import Plain
 
 
-class TestLicenseDefinitionData:
+class TestCopyrightNoticeDefinitionData:
     def test_summary(self) -> None:
         summary = Plain("My First Summary")
-        sut = LicenseDefinitionData(
+        sut = CopyrightNoticeDefinitionData(
             id="-dummy",
             label="-",
             summary=summary,
@@ -17,15 +17,17 @@ class TestLicenseDefinitionData:
 
     def test_text(self) -> None:
         text = Plain("My First Summary")
-        sut = LicenseDefinitionData(id="-dummy", label="-", summary="-", text=text)
+        sut = CopyrightNoticeDefinitionData(
+            id="-dummy", label="-", summary="-", text=text
+        )
         assert sut.text is text
 
     def test_new_plugin__minimal(self) -> None:
-        plugin_id = "my-first-license"
+        plugin_id = "my-first-copyright-notice"
         label = Plain("-")
         summary = Plain("-")
         text = Plain("-")
-        sut = LicenseDefinitionData(
+        sut = CopyrightNoticeDefinitionData(
             id=plugin_id,
             label=label,
             summary=summary,
@@ -39,8 +41,8 @@ class TestLicenseDefinitionData:
 
     def test_new_plugin__full(self) -> None:
         description = Plain("-")
-        sut = LicenseDefinitionData(
-            id="my-first-license",
+        sut = CopyrightNoticeDefinitionData(
+            id="my-first-copyright-notice",
             label="-",
             description=description,
             summary="-",
