+++ betty/tests/datas/plugin/definition/test_event_type.py
@@ -1,18 +1,19 @@
 from __future__ import annotations
 
-from betty.datas.gender_definition import GenderDefinitionData
+from betty.datas.plugin.definition.event_type import EventTypeDefinitionData
 from betty.localizables.plain import Plain
+from betty.machine_name import MachineName
 from betty.test_utils.locale.localizable import (
     DUMMY_COUNTABLE_LOCALIZABLE,
 )
 
 
-class TestGenderDefinitionData:
+class TestEventTypeDefinitionData:
     def test_new_plugin__minimal(self) -> None:
-        plugin_id = "my-first-gender"
+        plugin_id = "my-first-event-type"
         label = Plain("-")
         label_plural = Plain("-")
-        sut = GenderDefinitionData(
+        sut = EventTypeDefinitionData(
             id=plugin_id,
             label=label,
             label_plural=label_plural,
@@ -26,12 +27,18 @@ class TestGenderDefinitionData:
 
     def test_new_plugin__full(self) -> None:
         description = Plain("-")
-        sut = GenderDefinitionData(
-            id="my-first-gender",
+        before = MachineName("my-first-other-event-type")
+        after = MachineName("my-second-other-event-type")
+        sut = EventTypeDefinitionData(
+            id="my-first-event-type",
             label="-",
             label_plural="-",
             label_countable=DUMMY_COUNTABLE_LOCALIZABLE,
             description=description,
+            before={before},
+            after={after},
         )
         plugin = sut.new_plugin()
         assert plugin.description is description
+        assert plugin.before(before)
+        assert plugin.after(after)
