+++ betty/tests/datas/plugin/definition/test_place_type.py
@@ -1,19 +1,18 @@
 from __future__ import annotations
 
-from betty.datas.event_type_definition import EventTypeDefinitionData
+from betty.datas.plugin.definition.place_type import PlaceTypeDefinitionData
 from betty.localizables.plain import Plain
-from betty.machine_name import MachineName
 from betty.test_utils.locale.localizable import (
     DUMMY_COUNTABLE_LOCALIZABLE,
 )
 
 
-class TestEventTypeDefinitionData:
+class TestPlaceTypeDefinitionData:
     def test_new_plugin__minimal(self) -> None:
-        plugin_id = "my-first-event-type"
+        plugin_id = "my-first-place-type"
         label = Plain("-")
         label_plural = Plain("-")
-        sut = EventTypeDefinitionData(
+        sut = PlaceTypeDefinitionData(
             id=plugin_id,
             label=label,
             label_plural=label_plural,
@@ -27,18 +26,12 @@ class TestEventTypeDefinitionData:
 
     def test_new_plugin__full(self) -> None:
         description = Plain("-")
-        before = MachineName("my-first-other-event-type")
-        after = MachineName("my-second-other-event-type")
-        sut = EventTypeDefinitionData(
-            id="my-first-event-type",
+        sut = PlaceTypeDefinitionData(
+            id="my-first-place-type",
             label="-",
             label_plural="-",
             label_countable=DUMMY_COUNTABLE_LOCALIZABLE,
             description=description,
-            before={before},
-            after={after},
         )
         plugin = sut.new_plugin()
         assert plugin.description is description
-        assert plugin.before(before)
-        assert plugin.after(after)
